Restore status tests in previous position.
